requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
qdrant-client>=1.9.0
fastembed>=0.3.0
fastapi>=0.111.0
uvicorn>=0.30.0
pandas>=2.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
